fix latent integer issue in do_aes_decrypt by using explicit size_t arithmetic
This commit casts len1 and len2 explicitly to size_t before summing them into
out->len in do_aes_decrypt.

Although ciphertext_len is currently bounded and verified, performing addition on signed
int types for a size_t structure is a latent vulnerability that could lead to integer
overflow or signedness conversion issues under future refactoring of allocations.

+++ decryption.c
@@ -402,7 +402,7 @@ static int do_aes_decrypt(const uint8_t *in_data, size_t in_len,
 
     EVP_CIPHER_CTX_free(ctx);
 
-    out->len = len1 + len2;
+    out->len = (size_t)len1 + (size_t)len2;
     out->allocated_size = alloc_sz;
     out->buf = plain;
